+++ pipe_operator/__init__.py
@@ -1,5 +1,14 @@
 from .elixir_flow import elixir_pipe, tap, then
-from .python_flow import Pipe, PipeArgs, PipeEnd, PipeStart, Tap, Then
+from .python_flow import (
+    Pipe,
+    PipeArgs,
+    PipeEnd,
+    PipeStart,
+    Tap,
+    Then,
+    ThreadPipe,
+    ThreadWait,
+)
 
 __all__ = [
     "Pipe",
@@ -8,6 +17,8 @@ __all__ = [
     "PipeStart",
     "Tap",
     "Then",
+    "ThreadPipe",
+    "ThreadWait",
     "elixir_pipe",
     "tap",
     "then",